Updated: API, Web and database
API:
- [Added] BillService and BillController.
- [Changed] AccountController (added delivery api).
- [Added] CartAndCheckoutController.

Web:
- [Added] AdminBillController.
- [Added] CheckoutController.
- [Added] DeliveryController.
- [Added] PaymentMethodController.
- [Added] Copied model from api to web.
- [Changed] AccountRepository (copied from api).
- [Added] BillRepository.
- [Changed] CartRepository (copied from api).

Documents:
- [Changed] Draw.io for bumped version to 0.6.3.

+++ src/api/Model/Product.cs
@@ -25,7 +25,7 @@ namespace PhoneStoreManager.Model
 
         [Required]
         [JsonProperty("inventorycount")]
-        public int InventoryCount { get; set; } = 0;
+        public long InventoryCount { get; set; } = 0;
 
         [Required]
         [JsonProperty("unit")]
